requirements.txt
tweepy==4.14.0
google-api-python-client==2.147.0
google-auth-httplib2
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
